pyproject.toml
[tool.poetry]
name = "pichi"
version = "0.1.0"
description = ""
authors = ["Chris Land <chris.land@arcadia.io>"]

[tool.poetry.dependencies]
python = "^3.9"
auth0-python = "^3.14.0"
fastapi = "^0.63"
loguru = "^0.5"
uvicorn = "^0.13"
requests = "^2.25"
python-jose = {version = "^3.2.0", extras = ["cryptography"]}
boto3 = "^1.17.12"
fastapi-camelcase = "^1.0.2"
cachetools = "^4.2.1"
Cerberus = "^1.3.2"


[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^20.8b1"
flake8 = "^3.8.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
